add icons for editing and deleting cards

+++ src/components/RenderCards.vue
@@ -30,13 +30,54 @@
   <hr />
 
   <ul>
-    <li class="border border-gray-400" v-for="card in filteredMemoryCards" :key="card.cardHeader">
-      <span class="underline">{{ card.cardHeader }}</span>
-      <p>{{ card.cardText }}</p>
-      <span class="text-green-500">{{ card.category }}</span>
-      <span class="text-purple-500">{{ card.timestamp }}</span>
-      <span @click="deleteCard(card)" class="text-red-500">DELETE</span>
-      <span @click="editCard(card)" class="text-blue-500">EDIT</span>
+    <li
+      class="border border-gray-400 flex justify-between"
+      v-for="card in filteredMemoryCards"
+      :key="card.cardHeader"
+    >
+      <div>
+        <span class="underline">{{ card.cardHeader }}</span>
+        <p>{{ card.cardText }}</p>
+        <span class="text-green-500">{{ card.category }}</span>
+        <span class="text-purple-500">{{ card.timestamp }}</span>
+      </div>
+      <div class="flex-col">
+        <svg
+          xmlns="http://www.w3.org/2000/svg"
+          width="24"
+          height="24"
+          viewBox="0 0 24 24"
+          fill="none"
+          stroke="currentColor"
+          stroke-width="2"
+          stroke-linecap="round"
+          stroke-linejoin="round"
+          class="lucide lucide-trash-2"
+          @click="deleteCard(card)"
+        >
+          <path d="M3 6h18" />
+          <path d="M19 6v14c0 1-1 2-2 2H7c-1 0-2-1-2-2V6" />
+          <path d="M8 6V4c0-1 1-2 2-2h4c1 0 2 1 2 2v2" />
+          <line x1="10" x2="10" y1="11" y2="17" />
+          <line x1="14" x2="14" y1="11" y2="17" />
+        </svg>
+        <svg
+          xmlns="http://www.w3.org/2000/svg"
+          width="24"
+          height="24"
+          viewBox="0 0 24 24"
+          fill="none"
+          stroke="currentColor"
+          stroke-width="2"
+          stroke-linecap="round"
+          stroke-linejoin="round"
+          class="lucide lucide-pencil"
+          @click="editCard(card)"
+        >
+          <path d="M17 3a2.85 2.83 0 1 1 4 4L7.5 20.5 2 22l1.5-5.5Z" />
+          <path d="m15 5 4 4" />
+        </svg>
+      </div>
     </li>
   </ul>
 </template>
@@ -89,7 +130,13 @@ export default {
     },
 
     deleteCard(card) {
-      this.store.deleteCardToFirebase(card)
+      if (
+        window.confirm(
+          "Are you sure you want to delete this card? It can't be restored afterwards."
+        )
+      ) {
+        this.store.deleteCardToFirebase(card)
+      }
     },
 
     toggleAllNoneFilter(value) {
